chore: Add todo feature folder and organize
